Read the unload manifest and transfer only its entries
This makes sure that we don't accidentally transfer any other files if the prefix somehow includes files unrelated to the unload (this can happen if multiple unloads are performed to the same prefix from different clusters, or after a scale up or scale down).

The main motivation for this is that we want to have the manifest around to know exactly which files to remove when cleaning up, and it makes sense to use the manifest for the list of files to transfer when we need to load it anyway. It also means we don't have to explicitly exclude the manifest file.