Workaround Transfer Service's limit of 20 prefixes
The include_prefixes parameter of TransferSpec/ObjectConditions has a limit of 20 prefixes. This works when you only have 20 slices in your Redshift cluster, but not so well when you have more.

This changes the strategy to transfer the whole prefix, but to make it less likely that the prefix contains other files the prefix is now constructed using a file name prefix too, not just a directory structure.